move analyzer_rnn1_test out of analyzer_test
